[FLINK-10778] [tests] Move MigrationVersion to flink-core test utils
